requirements.txt
fastapi==0.103.1
uvicorn==0.22.0
Jinja2==3.1.2
itsdangerous>=2.1.2
python-multipart==0.0.20
aiogram==3.22.0
aiohttp==3.12.15
python_dotenv==1.2.1